+++ config/prelude.php
@@ -47,4 +47,6 @@ return [
     |
     */
     'timeout' => env('PRELUDE_TIMEOUT', 30),
+
+
 ];